requirements.txt
asetools==0.1.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
Pillow==10.0.1
pyparsing==3.2.3
pypng==0.0.20
python-dateutil==2.9.0.post0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.6.0+cu126 --find-links https://download.pytorch.org/whl/cu126
torchvision==0.21.0 --find-links https://download.pytorch.org/whl/cu126
triton==3.4.0
typing_extensions==4.14.1
